Refactor project structure into modularized architecture
Reorganized codebase into distinct `user-service`, `notification-service`, and `common` modules for improved modularity and reusability. Introduced package renaming and updated configurations to align with the new structure. Added shared common library for events and DTOs to streamline cross-module interactions.
