+++ common/src/main/java/com/wise/common/dto/CreateUserDto.java
@@ -1,4 +1,4 @@
-package com.wise.user.dto;
+package com.wise.common.dto;
 
 import lombok.Getter;
 import lombok.Setter;
