+++ common/src/main/java/com/wise/common/events/UserCreatedEvent.java
@@ -1,0 +1,15 @@
+package com.wise.common.events;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserCreatedEvent {
+    private Long id;
+    private String name;
+    private String email;
+    private String phone;
+}
