+++ common/src/main/resources/banner.txt
