+++ notification-service/src/main/java/com/wise/notification/NotificationServiceApplication.java
@@ -1,0 +1,11 @@
+package com.wise.notification;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class NotificationServiceApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(NotificationServiceApplication.class, args);
+    }
+}
