--- notification-service/src/main/java/com/wise/user/NotificationServiceApplication.java
@@ -1,11 +1,0 @@
-package com.wise.user;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class NotificationServiceApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(NotificationServiceApplication.class, args);
-    }
-}
