--- notification-service/src/main/java/com/wise/user/events/UserCreatedEvent.java
@@ -1,15 +1,0 @@
-package com.wise.user.events;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserCreatedEvent {
-    private Long id;
-    private String name;
-    private String email;
-    private String phone;
-}
