+++ user-service/src/main/java/com/wise/notification/UserServiceApplication.java
@@ -1,0 +1,11 @@
+package com.wise.notification;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class UserServiceApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(UserServiceApplication.class, args);
+    }
+}
