+++ user-service/src/main/java/com/wise/notification/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package com.wise.notification.repository;
+
+import com.wise.notification.model.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByEmail(String email);
+    Optional<User> findByPhone(String phone);
+}
