--- user-service/src/main/java/com/wise/user/config/OpenApiConfig.java
@@ -1,18 +1,0 @@
-package com.wise.user.config;
-
-import io.swagger.v3.oas.models.info.Info;
-import io.swagger.v3.oas.models.OpenAPI;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class OpenApiConfig {
-    @Bean
-    public OpenAPI userServiceOpenAPI() {
-        return new OpenAPI()
-                .info(new Info()
-                        .title("User Service API")
-                        .version("v1")
-                        .description("APIs for user management"));
-    }
-}
