--- user-service/src/main/java/com/wise/user/model/User.java
@@ -1,45 +1,0 @@
-package com.wise.user.model;
-
-import jakarta.persistence.*;
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Pattern;
-import lombok.*;
-
-import java.time.Instant;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Table(name = "users")
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    @NotBlank(message = "Name is mandatory")
-    private String name;
-
-    @Column(nullable = false, unique = true)
-    @Email(message = "Email must be a valid address containing '@' symbol.")
-    private String email;
-
-    @Pattern(regexp = "^\\+?[0-9]*$", message = "Phone number must contain only digits and optional leading '+'.")
-    @Column(unique = true)
-    private String phone;
-
-    @Column(nullable = false)
-    private String passwordHash;
-
-    @Column(nullable = false, updatable = false)
-    private Instant createdAt;
-
-    @PrePersist
-    public void prePersist() {
-        createdAt = Instant.now();
-    }
-}
