--- user-service/src/main/java/com/wise/user/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package com.wise.user.repository;
-
-import com.wise.user.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
-    Optional<User> findByPhone(String phone);
-}
